Move GC debugging bits to a separate file
